Fix de onboarding, que cargaba constantemente
